test: refactor hosted UI harness helpers
