+++ apps/api/src/__tests__/operator-ui.test.ts
@@ -1,6 +1,5 @@
 import test from "node:test";
 import assert from "node:assert/strict";
-import vm from "node:vm";
 
 import {
   operatorUiApprovalDecisionPath,
@@ -24,250 +23,12 @@ import {
   renderOperatorUiHtml,
   renderOperatorUiStyleCss,
 } from "../operator-ui.js";
+import { createHostedUiClientHarness, flushClientPromises } from "./operator-ui-harness.js";
 
 function assertContainsAll(body: string, patterns: RegExp[]): void {
   for (const pattern of patterns) {
     assert.match(body, pattern);
   }
-}
-
-type HostedUiFetchCall = {
-  path: string;
-  method: string;
-  body: unknown;
-};
-
-class FakeElement {
-  public value = "";
-  public textContent = "";
-  public className = "";
-  public disabled = false;
-  public hidden = false;
-  public isConnected = true;
-  public children: FakeElement[] = [];
-  private readonly listeners = new Map<string, Array<(event: { currentTarget: FakeElement }) => unknown>>();
-  private readonly descendants = new Map<string, FakeElement>();
-  private readonly attributes = new Map<string, string>();
-  private innerHtmlValue = "";
-
-  public get innerHTML(): string {
-    return this.innerHtmlValue;
-  }
-
-  public set innerHTML(value: string) {
-    this.innerHtmlValue = value;
-    this.descendants.clear();
-  }
-
-  public addEventListener(type: string, listener: (event: { currentTarget: FakeElement }) => unknown): void {
-    this.listeners.set(type, [...(this.listeners.get(type) ?? []), listener]);
-  }
-
-  public async click(): Promise<void> {
-    for (const listener of this.listeners.get("click") ?? []) {
-      await listener({ currentTarget: this });
-    }
-  }
-
-  public replaceChildren(...children: FakeElement[]): void {
-    this.children = children;
-  }
-
-  public append(child: FakeElement): void {
-    this.children.push(child);
-  }
-
-  public get firstElementChild(): FakeElement | undefined {
-    return this.children[0];
-  }
-
-  public remove(): void {
-    this.isConnected = false;
-  }
-
-  public querySelector(selector: string): FakeElement {
-    const existing = this.descendants.get(selector);
-    if (existing) {
-      this.syncAttributeFromInnerHtml(selector, existing);
-      return existing;
-    }
-    const created = new FakeElement();
-    this.syncAttributeFromInnerHtml(selector, created);
-    this.descendants.set(selector, created);
-    return created;
-  }
-
-  public querySelectorAll(selector: string): FakeElement[] {
-    if (selector === "[data-artifact-proposal]") {
-      return [this.querySelector(selector)];
-    }
-    return [];
-  }
-
-  public getAttribute(name: string): string | null {
-    return this.attributes.get(name) ?? null;
-  }
-
-  public setAttribute(name: string, value: string): void {
-    this.attributes.set(name, value);
-  }
-
-  private syncAttributeFromInnerHtml(selector: string, element: FakeElement): void {
-    const attributeName = selector.match(/^\[([^\]]+)\]$/)?.[1];
-    if (!attributeName) return;
-    const attributeValue = this.innerHTML.match(new RegExp(`${attributeName}="([^"]*)"`))?.[1] ?? "";
-    element.setAttribute(attributeName, attributeValue);
-  }
-}
-
-function createHostedUiClientHarness() {
-  const selectors = [
-    "#project-scope",
-    "#status",
-    "#tracks",
-    "#runs",
-    "#detail",
-    "#refresh",
-    "#project-create",
-    "#project-update",
-    "#track-create",
-    "#project-name",
-    "#project-repo-url",
-    "#project-local-repo-path",
-    "#project-workflow-policy",
-    "#project-planning-system",
-    "#track-title",
-    "#track-description",
-    "#track-priority",
-  ];
-  const elements = new Map(selectors.map((selector) => [selector, new FakeElement()]));
-  const scope = elements.get("#project-scope")!;
-  const trackPriority = elements.get("#track-priority")!;
-  trackPriority.value = "medium";
-
-  const projects = [{ id: "project-1", name: "Project One", repoUrl: "https://example.com/one", localRepoPath: "/repo/one", defaultWorkflowPolicy: "standard", defaultPlanningSystem: "native" }];
-  const tracks: Array<Record<string, unknown>> = [];
-  const runs: Array<Record<string, unknown>> = [];
-  const calls: HostedUiFetchCall[] = [];
-  let projectCounter = 2;
-  let trackCounter = 1;
-  let runCounter = 1;
-  let planningSessionId: string | undefined;
-
-  const document = {
-    querySelector(selector: string) {
-      const element = elements.get(selector);
-      if (!element) throw new Error(`Missing fake element for ${selector}`);
-      return element;
-    },
-    createElement() {
-      return new FakeElement();
-    },
-  };
-
-  async function fetch(path: string, init?: { method?: string; body?: string }) {
-    const method = init?.method ?? "GET";
-    const body = init?.body === undefined ? undefined : JSON.parse(init.body);
-    calls.push({ path, method, body });
-
-    if (path === "/projects" && method === "GET") {
-      return { ok: true, json: async () => ({ projects }) };
-    }
-    if (path === "/projects" && method === "POST") {
-      const project = { id: `project-${projectCounter++}`, ...(body as Record<string, unknown>) };
-      projects.push(project as typeof projects[number]);
-      return { ok: true, json: async () => ({ project }) };
-    }
-    if (path.startsWith("/projects/") && method === "PATCH") {
-      const projectId = decodeURIComponent(path.slice("/projects/".length));
-      const project = { id: projectId, ...(body as Record<string, unknown>) };
-      return { ok: true, json: async () => ({ project }) };
-    }
-    if (path.startsWith("/tracks?page=1&pageSize=20") && method === "GET") {
-      return { ok: true, json: async () => ({ tracks }) };
-    }
-    if (path === "/tracks" && method === "POST") {
-      const track = { id: `track-${trackCounter++}`, status: "new", ...(body as Record<string, unknown>) };
-      tracks.push(track);
-      return { ok: true, json: async () => ({ track }) };
-    }
-    if (/^\/tracks\/[^/]+$/.test(path) && method === "GET") {
-      const trackId = decodeURIComponent(path.slice("/tracks/".length));
-      const track = tracks.find((candidate) => candidate.id === trackId) ?? { id: trackId, title: trackId, status: "new" };
-      return { ok: true, json: async () => ({ track, planningContext: { planningSessionId }, artifacts: {} }) };
-    }
-    if (/^\/tracks\/[^/]+$/.test(path) && method === "PATCH") {
-      const trackId = decodeURIComponent(path.slice("/tracks/".length));
-      const track = { id: trackId, ...(body as Record<string, unknown>) };
-      return { ok: true, json: async () => ({ track }) };
-    }
-    if (/^\/tracks\/[^/]+\/planning-sessions$/.test(path) && method === "POST") {
-      planningSessionId = "planning-1";
-      return { ok: true, json: async () => ({ planningSession: { id: planningSessionId, ...(body as Record<string, unknown>) } }) };
-    }
-    if (/^\/planning-sessions\/[^/]+\/messages$/.test(path) && method === "POST") {
-      return { ok: true, json: async () => ({ message: { id: "message-1", ...(body as Record<string, unknown>) } }) };
-    }
-    if (/^\/tracks\/[^/]+\/artifacts\/(spec|plan|tasks)$/.test(path) && method === "GET") {
-      return { ok: true, json: async () => ({ approvalRequests: [] }) };
-    }
-    if (/^\/tracks\/[^/]+\/artifacts\/(spec|plan|tasks)$/.test(path) && method === "POST") {
-      return { ok: true, json: async () => ({ revision: { id: "revision-1", ...(body as Record<string, unknown>) } }) };
-    }
-    if (path === "/runs?page=1&pageSize=20" && method === "GET") {
-      return { ok: true, json: async () => ({ runs }) };
-    }
-    if (path === "/runs" && method === "POST") {
-      const run = { id: `run-${runCounter++}`, status: "running", ...(body as Record<string, unknown>) };
-      runs.push(run);
-      return { ok: true, json: async () => ({ run }) };
-    }
-    if (/^\/runs\/[^/]+$/.test(path) && method === "GET") {
-      const runId = decodeURIComponent(path.slice("/runs/".length));
-      const run = runs.find((candidate) => candidate.id === runId) ?? { id: runId, trackId: "track-1", status: "running" };
-      return { ok: true, json: async () => ({ run }) };
-    }
-    if (/^\/runs\/[^/]+\/events$/.test(path) && method === "GET") {
-      return { ok: true, json: async () => ({ events: [] }) };
-    }
-    if (/^\/runs\/[^/]+\/resume$/.test(path) && method === "POST") {
-      return { ok: true, json: async () => ({ run: { id: path.split("/")[2], status: "running", ...(body as Record<string, unknown>) } }) };
-    }
-    if (/^\/runs\/[^/]+\/cancel$/.test(path) && method === "POST") {
-      return { ok: true, json: async () => ({ run: { id: path.split("/")[2], status: "cancelled" } }) };
-    }
-    if (/^\/runs\/[^/]+\/workspace-cleanup\/preview$/.test(path) && method === "GET") {
-      return { ok: true, json: async () => ({ cleanupPlan: { eligible: true, operations: [{ kind: "delete", path: "/tmp/specrail-worktree" }], refusalReasons: [] } }) };
-    }
-    if (/^\/runs\/[^/]+\/workspace-cleanup\/apply$/.test(path) && method === "POST") {
-      if ((body as { confirm?: string }).confirm === "") {
-        return { ok: true, json: async () => ({ expectedConfirmation: "APPLY CLEANUP run-1" }) };
-      }
-      return { ok: true, json: async () => ({ cleanupResult: { status: "completed", failures: [] } }) };
-    }
-    throw new Error(`Unhandled fetch ${method} ${path}`);
-  }
-
-  vm.runInNewContext(renderOperatorUiClientScript(), {
-    document,
-    fetch,
-    Option: class FakeOption {
-      public textContent: string;
-      public value: string;
-      public constructor(label: string, value: string) {
-        this.textContent = label;
-        this.value = value;
-      }
-    },
-    EventSource: undefined,
-  });
-
-  return { calls, elements, scope };
-}
-
-async function flushClientPromises(): Promise<void> {
-  await new Promise((resolve) => setImmediate(resolve));
-  await new Promise((resolve) => setImmediate(resolve));
 }
 
 test("operator UI helpers escape metadata and previews", () => {
