+++ docs/architecture/mvp-roadmap.md
@@ -109,10 +109,11 @@ This roadmap reflects the implemented MVP baseline and the next practical gaps t
 - hosted UI tests guard against reintroducing browser-native prompt/confirm dialogs and group shell coverage by action area without adding a frontend build pipeline
 - hosted UI markup exposes stable `data-control-group` selectors for major action areas
 - hosted UI client action harness exercises top-level project/track, selected-track, and selected-run click handlers without a browser dependency
+- hosted UI fake DOM/fetch harness is isolated in a focused API test helper
 - keep HTTP/SSE as the system of record for new clients
 - reuse existing approval, event, and listing APIs rather than inventing parallel workflows
 
 ## Suggested issue framing from the current baseline
 
-1. **Hosted operator UI harness refactor**
-   - extract the growing no-dependency fake DOM/fetch harness into focused helpers so future UI action coverage stays maintainable.
+1. **Hosted operator UI harness ergonomics**
+   - add small helper methods for common hosted UI setup flows (create track/run, load cleanup preview) so future action tests stay compact.
